backend/requirements.txt
# All versions pinned - do not upgrade without testing

--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4
sentence-transformers==2.7.0
transformers==4.40.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
torch==2.2.2
numpy==1.26.4
yake==0.4.8